Overhaul `requirements` module
The `requirements` module worked, but was very unpythonic.
For this reason, I have changed how the requirement classes are
implemented.

The most important change is how the methods work together.
There used to be two methods, `is_compatible_with` and
`are_compatible_with`.
The former checked if one item fit the requirement,
and the latter called the former repeatedly to make sure all
requirements in `*args` fit the requirement.
However, this resulted in syntax that 1) was too wordy and 2)
made it difficult to understand how to use the requirement.
Now, a requirement has a "private" `_checker` and a "public"
`check(*args)` method.
The programmer implements the `_checker`, and the `check` calls the
other method. With this change, the code becomes much easier to read.

I also had to update existing `requirements` tests to utilize the
correct method names.

The second main change is the preferance of the `pydantic` package over
regular dataclasses.
Some of the requirement classes needed to use validation, which the
`pydantic` classes specialize in.
For this reason, I added the module here.

I decided to remove the `Requirement` tests because the only test in the
file tested if the class is abstract, which is not particularly
necessary and (more importantly) gets flagged as an issue in mypy.
With this change, I can change the tox configurations to check my tests
using mypy as well.

Some minor changes to existing code were needed to account for name
changes and the like.