Integrate Nikolaos' tools
I updated several parts of the `CTAInterface` class. First of all, I
made requirements into a variable that is simply set in each
implementation of the class to increase clarity and simplicity. Then, I
modified the outputs of the `convert` method to also allow for returning
a `pandas.DataFrame` to allow for more output variety. Since a
`DataFrame` cannot be comprised of `DataFrame` columns, I had to change
the output of the `annotate` function to a
`dict[str, Series | DataFrame]`

The celltypist interface needed some minor tweaks to increase
its functionality to at least match that of Nikolaos' code. Some minor
refactoring was also required to work with the modified requirements
module. I also added some more documentation for the sake of clarity.

I added one more extra/testenv called `torch38` housing the packages
that require Python 3.8 and torch. This is because torch and SCANVI take
 ~5GB for requirements for every virtual environment it is installed on,
 so it makes sense to only install it in one place.

 I created a `NotNoneRequirement` to require certain parameters to not
 be a `None` value.

 Also, I updated the versions of certain used GitHub actions.

 Finally, I ran `isort` on the `src` and `tests` directories to organize the
 downloads.