Create CLI
In this commit, I created the `macta-tools` CLI.

Additionally, I added another field to the `CTAToolInterface` class
called `_required_kwargs` - a list of strings that represent the names
of the kwargs that the tool must be passed, simplifying the syntax to do
so. Consequently, the `NotNoneRequirement` and its related files are no
longer needed, so I deleted them.

I also added improved the logging suppression in the SCANVI interface
file.

I added some small tweaks to the project configuration to re-add line
length checks and to configure where pyright searches for issues.